requirements.txt
asgiref==3.5.2
Django==4.1.3
django-cleanup==6.0.0
django-client-side-image-cropping==0.1.10
Pillow==9.3.0
sqlparse==0.4.3
tzdata==2022.6